serverutil/signup: move signup from serverutil/keyserver to new package
The signup handler needn't be part of the serverutil/keyserver package.
Isolating it will make it easier to test and maintain in the long term.

A step toward resolving issue #435.
