+++ serverutil/keyserver/main.go
@@ -8,8 +8,11 @@ package keyserver // import "upspin.io/serverutil/keyserver"
 
 import (
 	"flag"
+	"io/ioutil"
 	"net/http"
+	"strings"
 
+	"upspin.io/cloud/mail/sendgrid"
 	"upspin.io/cloud/storage"
 	"upspin.io/config"
 	"upspin.io/errors"
@@ -18,6 +21,7 @@ import (
 	"upspin.io/key/server"
 	"upspin.io/log"
 	"upspin.io/rpc/keyserver"
+	"upspin.io/serverutil/signup"
 	"upspin.io/upspin"
 
 	// Load required transports
@@ -72,10 +76,11 @@ func Main(setup func(upspin.KeyServer)) {
 	if logger, ok := key.(server.Logger); ok {
 		http.Handle("/log", logHandler{logger: logger})
 	}
+
 	if *mailConfigFile != "" {
 		f := cfg.Factotum()
 		if f == nil {
-			log.Fatal("supplied config must include keys when -mail_config set")
+			log.Fatal("keyserver: supplied config must include keys when -mail_config set")
 		}
 		project := ""
 		flag.Visit(func(f *flag.Flag) {
@@ -84,12 +89,28 @@ func Main(setup func(upspin.KeyServer)) {
 			}
 			project = f.Value.String()
 		})
-		h, err := newSignupHandler(f, key, *mailConfigFile, project)
+		apiKey, _, _, err := parseMailConfig(*mailConfigFile)
 		if err != nil {
-			log.Fatal(err)
+			log.Fatalf("keyserver: %v", err)
 		}
-		http.Handle("/signup", h)
+		m := sendgrid.New(apiKey, "upspin.io")
+		http.Handle("/signup", signup.NewHandler(f, key, m, project))
 	} else {
 		log.Println("keyserver: -mail_config not set, /signup deactivated")
 	}
+}
+
+func parseMailConfig(name string) (apiKey, userName, password string, err error) {
+	data, err := ioutil.ReadFile(name)
+	if err != nil {
+		return "", "", "", errors.E(errors.IO, err)
+	}
+	lines := strings.Split(strings.TrimSpace(string(data)), "\n")
+	if len(lines) != 3 {
+		return "", "", "", errors.E(errors.IO, errors.Str("config file must have 3 entries: api key, user name, password"))
+	}
+	apiKey = strings.TrimSpace(lines[0])
+	userName = strings.TrimSpace(lines[1])
+	password = strings.TrimSpace(lines[2])
+	return
 }
